move updated functionality to testing suite from knownDefects + refinements
